saslserv: use size_t in all places taking a parameter size
Conflicts:
	include/sasl.h
	modules/saslserv/dh-aes.c
	modules/saslserv/ecdsa-nist256p-challenge.c
	modules/saslserv/main.c

+++ modules/saslserv/main.c
@@ -265,7 +265,7 @@ static void sasl_packet(sasl_session_t *p, char *buf, int len)
 	char *cloak, *out = NULL;
 	char temp[BUFSIZE];
 	char mech[21];
-	int out_len = 0;
+	size_t out_len = 0;
 	metadata_t *md;
 
 	/* First piece of data in a session is the name of
